split up chart table
