+++ packages/ui-components/src/components/Chart/Legend.js
@@ -7,7 +7,6 @@ import React from 'react';
 import styled from 'styled-components';
 import MuiButton from '@material-ui/core/Button';
 import Tooltip from '@material-ui/core/Tooltip';
-import { formatDataValueByType } from '@tupaia/utils';
 import { isMobile } from './utils';
 
 const LegendContainer = styled.div`
@@ -52,8 +51,8 @@ const LegendItem = styled(({ isExporting, ...props }) => <MuiButton {...props} /
   // small styles
   &.small {
     font-size: 0.5rem;
-    padding-bottom: 0;
-    padding-top: 0;
+    padding-bottom: 0.1rem;
+    padding-top: 0.1rem;
     margin-right: 0;
   }
 `;
@@ -86,52 +85,28 @@ const Text = styled.span`
   line-height: 1.4;
 `;
 
-const getPieLegendDisplayValue = (chartConfig, value, item, isEnlarged, viewContent) => {
-  if (chartConfig[value]?.label) {
-    return chartConfig[value].label;
-  }
+const getDisplayValue = (chartConfig, value) => chartConfig[value]?.label || value;
 
-  const metadata = item[`${value}_metadata`];
-  const labelSuffix = formatDataValueByType({ value: item.value, metadata }, viewContent.valueType);
-
-  // on mobile the legend will show the actual formatDataValueByType after the label value
-  return isMobile() && isEnlarged ? `${value} ${labelSuffix}` : value;
-};
-
-export const getPieLegend = ({
-  chartConfig = {},
-  isEnlarged,
-  isExporting,
-  legendPosition,
-  viewContent,
-}) => ({ payload }) => (
+export const getPieLegend = ({ chartConfig = {}, isEnlarged, isExporting, legendPosition }) => ({
+  payload,
+}) => (
   <PieLegendContainer position={legendPosition}>
-    {payload.map(({ color, value, payload: item }) => {
-      const displayValue = getPieLegendDisplayValue(
-        chartConfig,
-        value,
-        item,
-        isEnlarged,
-        viewContent,
-      );
-
-      return (
-        <LegendItem
-          key={value}
-          isExporting={isExporting}
-          className={isEnlarged && !isMobile() ? 'enlarged' : 'small'}
-          disabled
-        >
-          <TooltipContainer>
-            <Box
-              className={isEnlarged && !isMobile() ? 'enlarged' : 'small'}
-              style={{ background: color }}
-            />
-            <Text>{displayValue}</Text>
-          </TooltipContainer>
-        </LegendItem>
-      );
-    })}
+    {payload.map(({ color, value }) => (
+      <LegendItem
+        key={value}
+        isExporting={isExporting}
+        className={isEnlarged && !isMobile() ? 'enlarged' : 'small'}
+        disabled
+      >
+        <TooltipContainer>
+          <Box
+            className={isEnlarged && !isMobile() ? 'enlarged' : 'small'}
+            style={{ background: color }}
+          />
+          <Text>{getDisplayValue(chartConfig, value)}</Text>
+        </TooltipContainer>
+      </LegendItem>
+    ))}
   </PieLegendContainer>
 );
 
@@ -144,8 +119,6 @@ export const getCartesianLegend = ({
 }) => ({ payload }) => (
   <LegendContainer position={legendPosition}>
     {payload.map(({ color, value, dataKey }) => {
-      const displayValue = chartConfig[value]?.label || value;
-
       return (
         <LegendItem
           key={value}
@@ -157,7 +130,7 @@ export const getCartesianLegend = ({
           <Tooltip title="Click to filter data" placement="top" arrow>
             <TooltipContainer>
               <Box className={isMobile() ? 'small' : 'enlarged'} style={{ background: color }} />
-              <Text>{displayValue}</Text>
+              <Text>{getDisplayValue(chartConfig, value)}</Text>
             </TooltipContainer>
           </Tooltip>
         </LegendItem>
