+++ packages/ui-components/src/components/Chart/Table/index.js
@@ -1,0 +1,1 @@
+export * from './Table';
