+++ packages/ui-components/src/components/Chart/index.js
@@ -4,7 +4,7 @@
  */
 
 export { Chart } from './Chart';
-export { Table } from './Table';
+export { Table } from './Table/Table';
 export * from './periodGranularities';
 export * from './parseChartConfig';
 export * from './constants';
